Posts
Most importantly - adding a login-form (as easy as possible to implement without the backend) (when you click Enter, user "login" and he becomes available all the information on the site. Added title page - greetings. Buttons to switch between posts and "About", as well as exit from the "account" are made in the navigation. Now you can see each post by opening it, see the comment and all the information about those who left it (with jsonplaceholder). Improved work with SPA, by routing the entire application. Also finalized the translation of the application into English.